ENH: Can now vertical markers locked to data

+++ applications/mne_analyze/extensions/rawdataviewer/fiffrawview.h
@@ -194,7 +194,7 @@ private:
 
     qint32                                              m_iT;                           /**< Display window size in seconds */
 
-    int                                              lastClickedPoint;
+    float                                              lastClickedPoint;
 signals:
     void tableViewDataWidthChanged(int iWidth);
 };
